+++ packages/shared/src/index.js
@@ -2,12 +2,15 @@ import * as TextHelpers from './TextHelpers'
 import * as DateTimeHelpers from './DateTimeHelpers/DateTimeHelpers'
 import * as Validators from './Validators'
 import * as LocationHelpers from './LocationHelpers'
+import * as OfferingHelpers from './OfferingHelpers'
 
 export * from './constants'
+export * from './OfferingHelpers'
 
 export {
   TextHelpers,
   DateTimeHelpers,
   Validators,
-  LocationHelpers
+  LocationHelpers,
+  OfferingHelpers
 }